Fix localWorkSize isn't less than or equal to globalWorkSize issue.
In the OpenCL, any value in localWorkSize should be less than the value
in globalWorkSize with the same index, what's more, it's dividable
by the value in globalWorkSize.

So take the Greatest Common Divisor of the value of globalWorkSize and
the value from |KERNEL_WORK_GROUP_SIZE| as the value in localWorkSize.