createBufferValidation, createImageValidation
When an image object is created as a CL object through createImage and read through readImage, the result is different from the original.
Change the image file to PNG and pass the test with the imageSmoothingEnabled option set to false.
It is assumed that the data changes during the interpolation process between lossy compression and canvas anti-alias.